+++ demo/src/create-proof-request.ts
@@ -1,0 +1,24 @@
+import { BigNumber, utils, Wallet } from 'ethers'
+import { Point } from '@noble/secp256k1'
+import { ProofRequest } from '@e2e-zk-ecdsa/shared'
+
+export const createProofRequest = async ({
+  addresses,
+  privateKey,
+}: {
+  addresses: string[]
+  privateKey: string
+}) => {
+  const wallet = new Wallet(privateKey)
+  const { address, publicKey } = wallet
+  const point = Point.fromHex(publicKey.slice(2))
+  const signature = await wallet.signMessage(1234n.toString())
+
+  return new ProofRequest({
+    addresses: addresses.map((address) => BigNumber.from(address).toBigInt()),
+    addressIndex: addresses.indexOf(address.toLowerCase()),
+    msghash: 1234n,
+    pubkey: point,
+    signature: utils.arrayify(signature),
+  })
+}
